// src/core/models/configuration.ts
export class Configuration {

    token: string;
    username: string;
    default_commit_message: string = "pippity poppity give me the zoppity";
    always_private: boolean = true;
    always_add_readme: boolean = true;
    always_add_license: boolean = false;
    always_add_gitignore: boolean = false;
}